% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMap.R
\name{createMap}
\alias{createMap}
\title{createMap}
\usage{
createMap(df, var, scale_title, plot_title, subtitle = "",
  low_color = "gray100", high_color = "gray0", scale_format = comma)
}
\arguments{
\item{df}{dataframe}

\item{var}{variable to plot}

\item{scale_title}{plot scale title}

\item{plot_title}{plot title}

\item{subtitle}{plot subtitle default is an empty string}

\item{low_color}{lightest color for mapping, default = 'gray100'}

\item{high_color}{darkest color for mapping, default = 'gray0'}

\item{scale_format}{changes the scale's formatting, default = comma (not scientific notation)}
}
\value{
returns a fiftystater map
}
\description{
Creates a new map using fiftystater based off of given dataframe, variable, title, and subtitle
}
\examples{
avg_per_state<-createMap(StateVsAvgCost, averageCost, "Average total cost", "Average cost of hospital procedures by state")

}
